Cập nhập sql và chức năng bán hàng

+++ QuanLyChuoiCuaHangTienLoi_Ministop/ET/ET_ChiTietHoaDon.cs
@@ -12,7 +12,7 @@ namespace ET
         public string MaHoaDon { get; set; }
         public string MaSanPham { get; set; }
         public int SoLuong { get; set; }
-        public double GiaBan { get; set; }
-        public double ThanhTien { get; set; }
+        public decimal GiaBan { get; set; }
+        public decimal ThanhTien { get; set; }
     }
 }
